feat #6 - Upcoming API 연결(Token 삭제함)
