+++ CinemaApp/Model/Login/UpcomingMovie.swift
@@ -1,0 +1,23 @@
+//
+//  UpcomingMovies.swift
+//  CinemaApp
+//
+//  Created by 유림 on 4/24/24.
+//
+
+import Foundation
+struct UpcomingMovie: Decodable {
+    let id: Int
+    let title: String
+    let posterPath: String
+    
+    public enum CodingKeys: String, CodingKey {
+        case id
+        case title
+        case posterPath = "poster_path"
+    }
+}
+
+struct UpcomingMovies: Decodable {
+    let results: [UpcomingMovie]
+}
