+++ CinemaApp/Presentation/login/StoryBoard/Login.storyboard
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Setup Login View Controller-->
+        <!--Login View Controller-->
         <scene sceneID="s0d-6b-0kx">
             <objects>
-                <viewController id="Y6W-OH-hqX" customClass="SetupLoginViewController" customModule="CinemaApp" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="Y6W-OH-hqX" customClass="LoginViewController" customModule="CinemaApp" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="5EZ-qb-Rvc">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -131,10 +131,10 @@
             </objects>
             <point key="canvasLocation" x="828.24427480916029" y="3.5211267605633805"/>
         </scene>
-        <!--Setup Register View Controller-->
+        <!--Register View Controller-->
         <scene sceneID="k83-Z2-phf">
             <objects>
-                <viewController id="aap-l5-SSd" customClass="SetupRegisterViewController" customModule="CinemaApp" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="aap-l5-SSd" customClass="RegisterViewController" customModule="CinemaApp" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="kag-aU-bHh">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="842"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
